Convert during/ensureAsync tests to mocha
