+++ test/test-async.js
@@ -156,165 +156,3 @@ exports['concatSeries'] = function(test){
         test.done();
     });
 };
-
-exports['during'] = function (test) {
-    var call_order = [];
-
-    var count = 0;
-    async.during(
-        function (cb) {
-            call_order.push(['test', count]);
-            cb(null, count < 5);
-        },
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb();
-        },
-        function (err) {
-            test.ok(err === null, err + " passed instead of 'null'");
-            test.same(call_order, [
-                ['test', 0],
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5],
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doDuring'] = function (test) {
-    var call_order = [];
-
-    var count = 0;
-    async.doDuring(
-        function (cb) {
-            call_order.push(['iteratee', count]);
-            count++;
-            cb();
-        },
-        function (cb) {
-            call_order.push(['test', count]);
-            cb(null, count < 5);
-        },
-        function (err) {
-            test.ok(err === null, err + " passed instead of 'null'");
-            test.same(call_order, [
-                ['iteratee', 0], ['test', 1],
-                ['iteratee', 1], ['test', 2],
-                ['iteratee', 2], ['test', 3],
-                ['iteratee', 3], ['test', 4],
-                ['iteratee', 4], ['test', 5],
-            ]);
-            test.equals(count, 5);
-            test.done();
-        }
-    );
-};
-
-exports['doDuring - error test'] = function (test) {
-    test.expect(1);
-    var error = new Error('asdas');
-
-    async.doDuring(
-        function (cb) {
-            cb(error);
-        },
-        function () {},
-        function (err) {
-            test.equal(err, error);
-            test.done();
-        }
-    );
-};
-
-exports['doDuring - error iteratee'] = function (test) {
-    test.expect(1);
-    var error = new Error('asdas');
-
-    async.doDuring(
-        function (cb) {
-            cb(null);
-        },
-        function (cb) {
-            cb(error);
-        },
-        function (err) {
-            test.equal(err, error);
-            test.done();
-        }
-    );
-};
-
-exports['ensureAsync'] = {
-    'defer sync functions': function (test) {
-        test.expect(6);
-        var sync = true;
-        async.ensureAsync(function (arg1, arg2, cb) {
-            test.equal(arg1, 1);
-            test.equal(arg2, 2);
-            cb(null, 4, 5);
-        })(1, 2, function (err, arg4, arg5) {
-            test.equal(err, null);
-            test.equal(arg4, 4);
-            test.equal(arg5, 5);
-            test.ok(!sync, 'callback called on same tick');
-            test.done();
-        });
-        sync = false;
-    },
-
-    'do not defer async functions': function (test) {
-        test.expect(6);
-        var sync = false;
-        async.ensureAsync(function (arg1, arg2, cb) {
-            test.equal(arg1, 1);
-            test.equal(arg2, 2);
-            async.setImmediate(function () {
-                sync = true;
-                cb(null, 4, 5);
-                sync = false;
-            });
-        })(1, 2, function (err, arg4, arg5) {
-            test.equal(err, null);
-            test.equal(arg4, 4);
-            test.equal(arg5, 5);
-            test.ok(sync, 'callback called on next tick');
-            test.done();
-        });
-    },
-
-    'double wrapping': function (test) {
-        test.expect(6);
-        var sync = true;
-        async.ensureAsync(async.ensureAsync(function (arg1, arg2, cb) {
-            test.equal(arg1, 1);
-            test.equal(arg2, 2);
-            cb(null, 4, 5);
-        }))(1, 2, function (err, arg4, arg5) {
-            test.equal(err, null);
-            test.equal(arg4, 4);
-            test.equal(arg5, 5);
-            test.ok(!sync, 'callback called on same tick');
-            test.done();
-        });
-        sync = false;
-    }
-};
-
-exports['constant'] = function (test) {
-    test.expect(5);
-    var f = async.constant(42, 1, 2, 3);
-    f(function (err, value, a, b, c) {
-        test.ok(!err);
-        test.ok(value === 42);
-        test.ok(a === 1);
-        test.ok(b === 2);
-        test.ok(c === 3);
-        test.done();
-    });
-};
